requirements.txt
websocket-client
redis
websockets
asyncio